Move powerstrip containers up to the root dir.
