+++ powerstrip/Dockerfile
