+++ powerstrip/root_overlay/etc/powerstrip/adapters.yml
