+++ powerstrip/root_overlay/etc/rc.local
