+++ powerstrip/root_overlay/etc/service/bird/run
