+++ powerstrip/root_overlay/etc/service/confd/run
