+++ powerstrip/root_overlay/etc/service/felix/run
