+++ powerstrip/root_overlay/etc/service/powerstrip-calico/run
