+++ powerstrip/root_overlay/etc/service/powerstrip/run
