+++ powerstrip_adapter/Dockerfile
