+++ powerstrip_adapter/root_overlay/adapter/calico_etcd.py
