+++ powerstrip_adapter/root_overlay/adapter/netns.py
