+++ powerstrip_adapter/root_overlay/adapter/powerstrip-calico.py
