+++ powerstrip_adapter/root_overlay/adapter/requirements.txt
